+++ R/webshot.R
@@ -1,7 +1,7 @@
 #' Embedded webshot of leaflet map
 #' @param map A leaflet map object
 #' @param path_to_img A path to the image file to save
-#' @param overwite Do you want to clobber any existing file?
+#' @param overwrite Do you want to clobber any existing file?
 #' @param vwidth see \code{\link[webshot]{webshot}}
 #' @param vheight see \code{\link[webshot]{webshot}}
 #' @param cliprect see \code{\link[webshot]{webshot}}
@@ -9,11 +9,13 @@
 #' @export
 #' @return a path to a PNG file
 #' @examples 
+#' \dontrun{
 #' if (require(leaflet)) {
 #'   map <- leaflet() %>%
 #'     addTiles() %>%
-#'     addMarkers(lng=174.768, lat=-36.852, popup="The birthplace of R")
+#'     addMarkers(lng = 174.768, lat = -36.852, popup = "The birthplace of R")
 #'   save_webshot(map, tempfile())
+#' }
 #' }
 
 save_webshot <- function(map, path_to_img, overwrite = FALSE,